backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
motor==3.6.0
pymongo==4.10.1

# Validation / Settings
pydantic==2.9.2
pydantic-settings==2.5.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.27.2

# AI / ML
sentence-transformers==3.1.1
chromadb==0.5.15
langchain==0.3.3
langchain-openai==0.2.3
langchain-community==0.3.2
openai==1.51.2

# Config
python-dotenv==1.0.1
PyYAML==6.0.2

# Utilities
python-multipart==0.0.12

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code quality
flake8==7.1.1
black==24.10.0
bandit==1.7.10
